Move html entry points into public
