+++ public/html/options.html
@@ -7,6 +7,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="/src/options/index.tsx" type="module"></script>
+<script src="/src/pages/options/index.tsx" type="module"></script>
 </body>
 </html>
