+++ public/html/popup.html
@@ -7,6 +7,6 @@
 </head>
 <body>
 <div id="root"></div>
-<script src="/src/popup/index.tsx" type="module"></script>
+<script src="/src/pages/popup/index.tsx" type="module"></script>
 </body>
 </html>
