--- src/options/index.html
@@ -1,12 +1,0 @@
-<!doctype html>
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <title>my ext option page</title>
-  </head>
-
-  <body>
-    <div id="my-ext-options-page"></div>
-    <script src="./index.tsx" type="module"></script>
-  </body>
-</html>
