+++ src/pages/popup/index.tsx
@@ -1,11 +1,11 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 
-import '../utils/fontAwesome';
+import '../../utils/fontAwesome';
 
 import Popup from './Popup';
 
-import '../index.css';
+import '../../index.css';
 
 ReactDOM.createRoot(document.getElementById('root') as HTMLElement).render(
   <React.StrictMode>
